% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.acceptfreq.timedeppar.R
\name{calc.acceptfreq}
\alias{calc.acceptfreq}
\title{Calculate apparent acceptance frequency of time-dependent parameter}
\usage{
calc.acceptfreq(x, n.burnin)
}
\arguments{
\item{x}{results from the function \code{infer.timedeppar} of class \code{timedeppar}.}

\item{n.burnin}{number of (unthinned) burnin points of the Markov chain to omit from analysis.}
}
\value{
list two-column matrices with time and apparent acceptance frequencies
}
\description{
This function calculates the apparent acceptance frequency from a potentially thinned Markov chain sample.
}
