% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.logpdf.timedeppar.R
\name{calc.logpdf}
\alias{calc.logpdf}
\title{Calculate log pdf values (prior, internal, posterior) from an object of type \code{timedeppar}}
\usage{
calc.logpdf(x, param, verbose)
}
\arguments{
\item{x}{results from the function \code{\link{infer.timedeppar}} of class \code{timedeppar}.}

\item{param}{list of parameter lists and vectors extracted from an object of class
\code{timedeppar} using the function \code{\link{get.param}}.}

\item{verbose}{boolean indicator for writing the results to the console
(default is not to do it).}
}
\value{
numeric vecctor of values of the different log pdfs.
}
\description{
This function calculated log priors, log pdf of Ornstein-Uhlenbeck time dependent parameters, and
log posterior from an object of type \code{timedeppar} produced by the function \code{\link{infer.timedeppar}}.
}
