% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.label.R
\name{get.label}
\alias{get.label}
\title{Construct a plot label from expressions for variables and units}
\usage{
get.label(var.name, labels = NA, units = NA, t1 = "", t2 = "", t3 = "")
}
\arguments{
\item{var.name}{name of the variable.}

\item{labels}{named vector of expressions encoding variable names with greek letters, sub- and superscripts.}

\item{units}{named vector of expressions encoding units with sub- and superscripts.}

\item{t1}{optional text (see below).}

\item{t2}{optional text (see below).}

\item{t3}{optional text (see below).}
}
\value{
expression of a label of the form: t1 label t2 [unit] t3\cr
        label and unit are extracted from the vectors \code{labels} and \code{units}
        by using the compontents named \code{var.name}
}
\description{
This function produces an expression to label plots from expressions for variables and units and from character strings.
}
