% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readres.timedeppar.R
\name{readres.timedeppar}
\alias{readres.timedeppar}
\title{Reads an object of type \code{timedeppar} saved to a file by \code{\link{infer.timedeppar}}}
\usage{
readres.timedeppar(file)
}
\arguments{
\item{file}{file name of the workspace to be read.}
}
\value{
object of type \code{timedeppar} containing the intermediate or final results of the inference process 
        or a list of length zero if the file was not found of no object called \code{res} of type \code{timedeppar}
        was found in the workspace.
}
\description{
This function read a workspace stored by the function \code{\link{infer.timedeppar}} and returns
the object of type \code{timedeppar} that contains the intermediate or final results of the 
inference process
}
